Updated versions of dependencies and broke out all grunt tasks into separate files
