+++ .travis.yml
@@ -1,6 +1,6 @@
 language: node_js
 node_js:
-  - '0.10'
+  - '5'
 before_install:
   - currentfolder=${PWD##*/}
   - if [ "$currentfolder" != 'generator-libmc' ]; then cd .. && eval "mv $currentfolder generator-libmc" && cd generator-libmc; fi
