--- app/templates/_bower.json
@@ -1,6 +1,0 @@
-{
-  "name": "package",
-  "version": "0.0.0",
-  "dependencies": {}
-}
-
