+++ app/templates/_package.json
@@ -8,24 +8,25 @@
     "test": "grunt test"
   },
   "devDependencies": {
-    "babel-bits": "^0.3.0",
-    "bit-imports": "^0.3.6",
-    "browserify": "~11.0.1",
-    "chai": "~3.2.0",
-    "eslint-tap": "^1.0.0",
+    "babel-bits": "^1.0.1",
+    "bit-imports": "^2.0.1",
+    "browserify": "~13.0.0",
+    "chai": "~3.5.0",
+    "config-grunt-tasks": "^1.0.0",
+    "eslint-tap": "^1.1.0",
     "grunt": "~0.4.5",
-    "grunt-browserify": "~4.0.1",
-    "grunt-concurrent": "~2.0.3",
-    "grunt-contrib-connect": "~0.11.2",
-    "grunt-contrib-uglify": "~0.9.2",
-    "grunt-contrib-watch": "~0.6.1",
-    "grunt-eslint": "^17.1.0",
-    "grunt-mocha": "~0.4.13",
+    "grunt-browserify": "~5.0.0",
+    "grunt-concurrent": "~2.2.1",
+    "grunt-contrib-connect": "~1.0.1",
+    "grunt-contrib-uglify": "~1.0.1",
+    "grunt-contrib-watch": "~1.0.0",
+    "grunt-eslint": "^18.0.0",
+    "grunt-mocha": "~0.4.15",
     "grunt-release": "^0.13.0",
-    "load-grunt-tasks": "^3.2.0",
-    "mocha": "~2.3.0",
-    "sinon": "~1.16.1",
-    "spromise": "^0.9.3"
+    "load-grunt-tasks": "^3.4.1",
+    "mocha": "~2.4.5",
+    "sinon": "~1.17.3",
+    "spromise": "^1.0.0"
   },
   "browser": {},
   "directories": {
