+++ app/templates/_travis.yml
@@ -1,6 +1,6 @@
 language: node_js
 node_js:
-  - "0.12"
+  - "5"
 
 before_install:
   - npm install -g grunt-cli
