+++ app/templates/tasks/concurrent.js
@@ -1,0 +1,8 @@
+module.exports = {
+  build: {
+    tasks: ['connect:keepalive', 'watch:build'],
+    options: {
+      logConcurrentOutput : true
+    }
+  }
+};
