+++ app/templates/tasks/eslint.js
@@ -1,0 +1,8 @@
+module.exports = {
+  all: {
+    options: {
+      //format: require('eslint-tap')
+    },
+    src: ['src/**/*.js', 'test/**/*.js', '*.js']
+  }
+};
