+++ app/templates/tasks/release.js
@@ -1,0 +1,8 @@
+module.exports = {
+  options: {
+    tagName: 'v<%= version %>',
+    tagMessage: 'Version <%= version %>',
+    commitMessage: 'Release v<%= version %>',
+    afterBump: ['build']
+  }
+};
