+++ app/templates/tasks/uglify.js
@@ -1,0 +1,11 @@
+module.exports = {
+  build: {
+    options: {
+      preserveComments: 'some',
+      sourceMap: true
+    },
+    files: {
+      'dist/index.min.js': ['dist/index.js']
+    }
+  }
+};
