+++ app/templates/tasks/watch.js
@@ -1,0 +1,11 @@
+var env = require("./env");
+
+module.exports = {
+  build: {
+    files: ['src/**/*.js', 'test/**/*.js', '*.js'],
+    tasks: ['build'],
+    options: {
+      livereload: env.livereloadPortNumber
+    }
+  }
+};
