+++ package.json
@@ -1,7 +1,7 @@
 {
   "name": "generator-libmc",
   "version": "2.0.0",
-  "description": "Yeoman generator for scaffolding micro libraries that run in the Browser and Node.js. It has browserify, mocha, chai, sinon, phantomjs, connect, jshint, bit-imports, and grunt as the core stack for linting, building, and testing.",
+  "description": "Yeoman generator for scaffolding micro libraries that run in the Browser and Node.js. It has browserify, mocha, chai, sinon, phantomjs, connect, eslint, bit-imports, and grunt as the core stack for linting, building, and testing.",
   "license": "MIT",
   "main": "app/index.js",
   "repository": "https://github.com/MiguelCastillo/generator-libmc",
@@ -28,9 +28,9 @@
     "sinon"
   ],
   "dependencies": {
-    "yeoman-generator": "^0.18.0",
-    "chalk": "^0.5.0",
-    "yosay": "^0.3.0"
+    "yeoman-generator": "^0.22.5",
+    "chalk": "^1.1.1",
+    "yosay": "^1.1.0"
   },
   "devDependencies": {
     "mocha": "*"
