+++ test/test-app.js
@@ -18,7 +18,6 @@ describe('libmc:app', function () {
 
   it('creates files', function () {
     assert.file([
-      'bower.json',
       'package.json',
       '.editorconfig',
       '.jshintrc'
